+++ GjBaseLibrary/build.gradle
@@ -1,16 +1,14 @@
 apply plugin: 'com.android.library'
 
-def depend = rootProject.ext.dependencies
-
 android {
-    compileSdkVersion COMPILE_SDK_VERSION as int
-    buildToolsVersion BUILD_TOOLS_VERSION as String
+    compileSdkVersion rootProject.ext.android.compileSdkVersion
+    buildToolsVersion rootProject.ext.android.buildToolsVersion
 
     defaultConfig {
-        minSdkVersion MIN_SDK_VERSION as int
-        targetSdkVersion TARGET_SDK_VERSION as int
-        versionCode VERSION_CODE as int
-        versionName VERSION_NAME as String
+        minSdkVersion rootProject.ext.android.minSdkVersion
+        targetSdkVersion rootProject.ext.android.targetSdkVersion
+        versionCode rootProject.ext.android.versionCode
+        versionName rootProject.ext.android.versionName
     }
     buildTypes {
         release {
@@ -18,12 +16,29 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
     compile fileTree(include: ['*.jar'], dir: 'libs')
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.3.0'
-    compile 'com.orhanobut:logger:2.1.1'
-    compile depend.support
-    compile depend.glide
+
+    compile rootProject.ext.dependencies.support
+    compile rootProject.ext.dependencies.rxjava
+    compile rootProject.ext.dependencies.okhttp
+    compile rootProject.ext.dependencies.glide
+    compile rootProject.ext.dependencies.retrofit
+    compile rootProject.ext.dependencies.logger
+    compile rootProject.ext.dependencies.cookie
+    compile rootProject.ext.dependencies.diskLruCache
+
+
+//    compile rootProject.ext.dependencies.luban
+//    compile rootProject.ext.dependencies.rxbinding
+//    compile(rootProject.ext.dependencies.multitype, {
+//        exclude group: 'com.android.support'
+//    })
 }
