+++ app/src/main/AndroidManifest.xml
@@ -3,7 +3,8 @@
           package="tv.guojiang.base">
 
     <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"></uses-permission>
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
     <application
         android:name=".App"
         android:allowBackup="true"
@@ -20,6 +21,13 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+
+        <activity android:name="tv.guojiang.sample.LogSampleActivity"/>
+
+        <activity android:name="tv.guojiang.sample.ImageLoadSampleActivity"/>
+
+        <activity android:name="tv.guojiang.sample.NetworkSampleActivity"/>
+
     </application>
 
 </manifest>
