--- app/src/main/java/tv/guojiang/LogSampleActivity.java
@@ -1,8 +1,0 @@
-package tv.guojiang;
-
-/**
- * @author leo
- */
-public class LogSampleActivity {
-
-}
