+++ build.gradle
@@ -1,5 +1,5 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
 apply from: 'config.gradle'
+
 buildscript {
     repositories {
         jcenter()
@@ -14,6 +14,7 @@ allprojects {
     repositories {
         jcenter()
         google()
+        maven { url "https://jitpack.io" }
         maven { url "https://dl.google.com/dl/android/maven2/" }
     }
 }
